requirements.txt
streamlit==1.24.0
streamlit-drawable-canvas==0.9.2
torch==2.0.1
numpy==1.24.3
Pillow==9.5.0
psycopg2-binary==2.9.6